Censor application name in pg_stat_activity from unprivileged users
This patch gives application_name the same treatment as query
in pg_stat_activity so that only a superuser or the current_user
can view their content.

This is important in multi-tenant systems where it is possible
for application_name to accidentally leak information about what
process is currently running, e.g. when the underlying postgres
driver utilizes the current process name ($0) to set a default
application name. Even if this were not the case, developer-set
application names should not be visible to other database users.